Test results displaying on spec/notesSpec.html, isLessThan and isGreaterThan have been added to the test framework, note.js created and content variable is working (test passing)
