+++ spec/notesSpec.html
@@ -1,0 +1,32 @@
+<html lang="en" dir="ltr">
+  <head>
+    <meta charset="utf-8">
+    <title></title>
+  </head>
+  <body>
+    <div id="results">
+      </div>
+    <script src='../bucket-test-framework/bucket.js'></script>
+    <script src='../src/note.js'></script>
+    <script src='./noteSpec.js'></script>
+    <script>
+      // const createDouble = bucket.createDouble;
+      // const describe = bucket.describe;
+      // const it = bucket.it;
+      // const expect = bucket.expect;
+      // const Note = require('../src/note.js');
+
+      // describe('Note', function (){
+      //   it('has text content', function () {
+      //     var note1 = new Note('This is a new note');
+      //     expect(note1.content).toBe('This is a new note');
+      //   });
+      // });
+
+      console.log("Total tests: " + (passes + fails));
+      console.log("Passes: " + (passes));
+      console.log("Fails: " + (fails));
+    </script>
+
+  </body>
+</html>
