+++ src/note.js
@@ -1,0 +1,3 @@
+function Note(content) {
+  this.content = content;
+}
